# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "KcomApi to communicate with TrackerServer"
host: "167.99.146.9:8082"
basePath: "/api"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- basicAuth: []
paths:
  /commands:
    get:
      summary: "Fetch a list of Saved Commands"
      description: "Without params, it returns a list of Drivers the user has access\
        \ to"
      operationId: "commandsGET"
      parameters:
      - name: "all"
        in: "query"
        description: "Can only be used by admins or managers to fetch all entities"
        required: false
        type: "boolean"
      - name: "userId"
        in: "query"
        description: "Standard users can use this only with their own _userId_"
        required: false
        type: "integer"
      - name: "deviceId"
        in: "query"
        description: "Standard users can use this only with _deviceId_s, they have\
          \ access to"
        required: false
        type: "integer"
      - name: "groupId"
        in: "query"
        description: "Standard users can use this only with _groupId_s, they have\
          \ access to"
        required: false
        type: "integer"
      - name: "refresh"
        in: "query"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Command"
      x-swagger-router-controller: "Default"
    post:
      summary: "Create a Saved Command"
      operationId: "commandsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Command"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Command"
      x-swagger-router-controller: "Default"
  /commands/{id}:
    put:
      summary: "Update a Saved Command"
      operationId: "commandsIdPUT"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Command"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Command"
      x-swagger-router-controller: "Default"
    delete:
      summary: "Delete a Saved Command"
      operationId: "commandsIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "Default"
  /commands/send:
    get:
      summary: "Fetch a list of Saved Commands supported by Device at the moment"
      description: "Return a list of saved commands linked to Device and its groups,\
        \ filtered by current Device protocol support"
      operationId: "commandsSendGET"
      parameters:
      - name: "deviceId"
        in: "query"
        description: "Standard users can use this only with _deviceId_s, they have\
          \ access to"
        required: false
        type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Command"
        "400":
          description: "Could happen when the user doesn't have permission for the\
            \ device"
      x-swagger-router-controller: "Default"
    post:
      summary: "Dispatch commands to device"
      description: "Dispatch a new command or Saved Command if _body.id_ set"
      operationId: "commandsSendPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Command"
      responses:
        "200":
          description: "Command sent"
          schema:
            $ref: "#/definitions/Command"
        "202":
          description: "Command queued"
          schema:
            $ref: "#/definitions/Command"
        "400":
          description: "Could happen when the user doesn't have permission or an incorrect\
            \ command _type_ for the device"
      x-swagger-router-controller: "Default"
  /commands/types:
    get:
      summary: "Fetch a list of available Commands for the Device or all possible\
        \ Commands if Device ommited"
      operationId: "commandsTypesGET"
      parameters:
      - name: "deviceId"
        in: "query"
        description: "Internal device identifier. Only works if device has already\
          \ reported some locations"
        required: false
        type: "integer"
      - name: "protocol"
        in: "query"
        description: "Protocol name. Can be used instead of device id"
        required: false
        type: "string"
      - name: "textChannel"
        in: "query"
        description: "When `true` return SMS commands. If not specified or `false`\
          \ return data commands"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CommandType"
        "400":
          description: "Could happen when trying to fetch from a device the user does\
            \ not have permission"
      x-swagger-router-controller: "Default"
  /devices:
    get:
      summary: "Fetch a list of Devices"
      description: "Without any params, returns a list of the user's devices"
      operationId: "devicesGET"
      parameters:
      - name: "all"
        in: "query"
        description: "Can only be used by admins or managers to fetch all entities"
        required: false
        type: "boolean"
      - name: "userId"
        in: "query"
        description: "Standard users can use this only with their own _userId_"
        required: false
        type: "integer"
      - name: "id"
        in: "query"
        description: "To fetch one or more devices. Multiple params can be passed\
          \ like `id=31&id=42`"
        required: false
        type: "integer"
        collectionFormat: "multi"
      - name: "uniqueId"
        in: "query"
        description: "To fetch one or more devices. Multiple params can be passed\
          \ like `uniqueId=333331&uniqieId=44442`"
        required: false
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Device"
        "400":
          description: "No permission"
      x-swagger-router-controller: "Default"
    post:
      summary: "Create a Device"
      operationId: "devicesPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Device"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Device"
      x-swagger-router-controller: "Default"
  /devices/{id}:
    put:
      summary: "Update a Device"
      operationId: "devicesIdPUT"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Device"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Device"
      x-swagger-router-controller: "Default"
    delete:
      summary: "Delete a Device"
      operationId: "devicesIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "Default"
  /devices/{id}/accumulators:
    put:
      summary: "Update total distance and hours of the Device"
      operationId: "devicesIdAccumulatorsPUT"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/DeviceAccumulators"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "Default"
  /groups:
    get:
      summary: "Fetch a list of Groups"
      description: "Without any params, returns a list of the Groups the user belongs\
        \ to"
      operationId: "groupsGET"
      parameters:
      - name: "all"
        in: "query"
        description: "Can only be used by admins or managers to fetch all entities"
        required: false
        type: "boolean"
      - name: "userId"
        in: "query"
        description: "Standard users can use this only with their own _userId_"
        required: false
        type: "integer"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
    post:
      summary: "Create a Group"
      operationId: "groupsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Group"
        "400":
          description: "No permission"
      x-swagger-router-controller: "Default"
  /groups/{id}:
    put:
      summary: "Update a Group"
      operationId: "groupsIdPUT"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
    delete:
      summary: "Delete a Group"
      operationId: "groupsIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "Default"
  /permissions:
    post:
      summary: "Link an Object to another Object"
      operationId: "permissionsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Permission"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Permission"
        "400":
          description: "No permission"
      x-swagger-router-controller: "Default"
    delete:
      summary: "Unlink an Object from another Object"
      operationId: "permissionsDELETE"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Permission"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "Default"
...
  /server:
    get:
      summary: "Fetch Server information"
      operationId: "serverGET"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Server"
      x-swagger-router-controller: "Default"
    put:
      summary: "Update Server information"
      operationId: "serverPUT"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Server"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/Server"
      x-swagger-router-controller: "Default"
...
securityDefinitions:
  basicAuth:
    description: "Basic HTTP authorization with _email_ and _password_"
    type: "basic"
definitions:
...
  Command:
    properties:
      id:
        type: "integer"
      deviceId:
        type: "integer"
      description:
        type: "string"
      type:
        type: "string"
      attributes:
        type: "object"
        properties: {}
    example:
      description: "description"
      attributes: "{}"
      id: 0
      type: "type"
      deviceId: 6
...
  Group:
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      groupId:
        type: "integer"
      attributes:
        type: "object"
        properties: {}
    example:
      groupId: 6
      name: "name"
      attributes: "{}"
      id: 0
...
  CommandType:
    properties:
      type:
        type: "string"
    example:
      type: "type"
...
  DeviceAccumulators:
    properties:
      deviceId:
        type: "integer"
      totalDistance:
        type: "number"
        description: "in meters"
      hours:
        type: "number"
    example:
      hours: 1.46581298050294517310021547018550336360931396484375
      totalDistance: 6.02745618307040320615897144307382404804229736328125
      deviceId: 0
...